solving compiler warnings

+++ bigint_arithmetic_test_cmocka.c
@@ -14,7 +14,7 @@ int tear_down(void** state){
     return 0;
 }
 
-void test_runner(uint64_t u[], uint64_t v[], uint64_t result[], uint64_t expected[], int num_operands_digits, int num_result_digits, uint64_t* (*func_to_test)(int, uint64_t*, uint64_t*, uint64_t*)){
+void test_runner(uint64_t u[], uint64_t v[], uint64_t result[], uint64_t expected[], size_t num_operands_digits, size_t num_result_digits, uint64_t* (*func_to_test)(int, uint64_t*, uint64_t*, uint64_t*)){
     (*func_to_test)(num_operands_digits, u, v, result);
     for (size_t i = 0; i < num_result_digits; i++)
         assert_int_equal(result[i], expected[i]);
@@ -24,7 +24,7 @@ void sum_trivial_case(void** state)
 {
     uint64_t u[] = {1};
     uint64_t v[] = {2};
-    int num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
+    size_t num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
     uint64_t result[num_digits+1];
     uint64_t expected[] = {3, 0};
     test_runner(u, v, result, expected, num_digits, num_digits+1, sum);
@@ -42,7 +42,7 @@ void sum_overflow(void** state) {
 void subtract_positive_result(void** state){
     uint64_t u[] = {3};
     uint64_t v[] = {1};
-    int num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
+    size_t num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
     uint64_t result[num_digits];
     uint64_t expected[] = {2};
     test_runner(u, v, result, expected, num_digits, num_digits, subtract);
@@ -51,7 +51,7 @@ void subtract_positive_result(void** state){
 void subtract_zero_result(void** state){
     uint64_t u[] = {3};
     uint64_t v[] = {3};
-    int num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
+    size_t num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
     uint64_t result[num_digits];
     uint64_t expected[] = {0};
     test_runner(u, v, result, expected, num_digits, num_digits, subtract);
@@ -60,7 +60,7 @@ void subtract_zero_result(void** state){
 void subtract_multiple_digits(void** state){
     uint64_t u[] = {123456789012345678lu, 576839484849498888lu};
     uint64_t v[] = {987654321098765432lu, 174857487482847488lu};
-    int num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
+    size_t num_digits = fmax(sizeof(u)/sizeof(u[0]), sizeof(v)/sizeof(v[0]));
     uint64_t result[num_digits];
     uint64_t expected[] = {135802467913580246, 401981997366651399};
     test_runner(u, v, result, expected, num_digits, num_digits, subtract);
